chore: split up record_spec
